+++ lib/openapi_parser/spec_validator.rb
@@ -25,6 +25,7 @@ require_relative 'spec_validator/rules/xml_attribute_deprecation'
 require_relative 'spec_validator/rules/xml_wrapped_deprecation'
 require_relative 'spec_validator/rules/security_scheme_fields_before_32'
 require_relative 'spec_validator/rules/media_types_before_32'
+require_relative 'spec_validator/rules/streaming_fields_before_32'
 
 module OpenAPIParser
   class SpecViolationError < OpenAPIError
@@ -97,6 +98,7 @@ module OpenAPIParser
           Rules::XmlWrappedDeprecation,
           Rules::SecuritySchemeFieldsBefore32,
           Rules::MediaTypesBefore32,
+          Rules::StreamingFieldsBefore32,
         ]
       end
   end